+++ pkg/ide/types.go
@@ -38,7 +38,7 @@ func (o Options) GetValue(values map[string]config.OptionValue, key string) stri
 	return ""
 }
 
-// ReusesAuthSock determines if the --reuse-sock flag should be passed to the ssh server helper based on the IDE.
+// ReusesAuthSock determines if the --reuse-ssh-auth-sock flag should be passed to the ssh server helper based on the IDE.
 // Browser based IDEs use a browser tunnel to communicate with the remote server instead of an independent ssh connection
 func ReusesAuthSock(ide string) bool {
 	return ide == "openvscode" || ide == "marimo" || ide == "jupyternotebook" || ide == "jlab"
